Make clear system evolution defines delta of state
In equation (3) in BluSTL_tutorial.pdf, the equation suggests that
the discrete form of the system evolution defines the next state
as a whole. In reality, the system evolution defines the change
in the system state. The updated documentation has not been built
into the PDF due to mild formatting issues with how the screenshots
are integrated into the PDF.